resolve some  conflict on 3d extension branch and sync develop to 3d branch

+++ cocos/2d/CCAtlasNode.cpp
@@ -263,12 +263,12 @@ TextureAtlas * AtlasNode::getTextureAtlas() const
     return _textureAtlas;
 }
 
-int AtlasNode::getQuadsToDraw() const
+ssize_t AtlasNode::getQuadsToDraw() const
 {
     return _quadsToDraw;
 }
 
-void AtlasNode::setQuadsToDraw(int quadsToDraw)
+void AtlasNode::setQuadsToDraw(ssize_t quadsToDraw)
 {
     _quadsToDraw = quadsToDraw;
 }
